coinTax/archive/Lighter 복사본/requirements.txt
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=0.19.0
